pyproject.toml
[project]
name = "coco-segmentation-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "pycocotools>=2.0.8",
    "tqdm>=4.67.1",
]
